Add functionality to add multiple participants to existing thread at once
